prometheus_generation/requirements.txt
# FastAPI и веб-сервер
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Валидация данных
pydantic>=2.0.0

# YAML парсер
pyyaml>=6.0

# S3
boto3



